fix(hydration-error-command-menu)
