+++ apps/next/src/app/(app)/page.tsx
@@ -1,7 +1,6 @@
 'use client'
 import { motion } from 'framer-motion'
-import dynamic from 'next/dynamic'
-// import CommandMenu from '~/components/CommandMenu'
+import CommandMenu from '~/components/CommandMenu'
 import Footer from '~/components/Footer'
 import Header from '~/components/Header'
 import { SessionCompleteModal } from '~/components/SessionCompleteModal'
@@ -15,8 +14,6 @@ export default function Dashboard() {
 
   const { currentSession, sessions } = useTimerStore()
   console.log('Dashboard store state:', currentSession, sessions)
-  const CommandMenu = dynamic(() => import('~/components/CommandMenu'), { ssr: false })
-
 
   return (
     <div>
