+++ apps/next/src/components/helper/ClientOnly.tsx
@@ -15,7 +15,7 @@ export function ClientOnly({ children, fallback = null }: ClientOnlyProps) {
   }, [])
 
   if (!hasMounted) {
-    return <>{fallback}</>
+    return fallback
   }
 
   return <>{children}</>
